fix: add query params hook to update query params and state simutaneously
